fix(numeric): 🐛 close IEEE 754 audit and fix i16 wrapping mul UB — Task 14 Tier A complete
## Summary

Close the Task 14 Tier A IEEE 754 audit by verifying full compliance
across f64 and f32 codegen, fixing a real UB bug in narrow wrapping
arithmetic, adding runtime-level tests, and updating contract/plan
status to reflect audit closure.

## Highlights

- Fix UB in __dao_wrapping_mul_i16: (uint16_t)a * (uint16_t)b promotes
  to signed int via C integer promotion; 65535*65535 overflows int.
  Fixed by casting to uint32_t before multiplication.
- Add 26 runtime-level overflow tests (wrapping + saturating for all
  signed widths) including the exact boundary case that triggered UB.
- Add f32 IEEE 754 comparison predicate tests (oeq, une, olt) to match
  existing f64 coverage.
- Update CONTRACT_NUMERIC_SEMANTICS status matrix: f64 IEEE 754,
  NaN/Inf/signed-zero, and comparison partial-order → Implemented.
- Mark Task 14 Tier A complete and exit criteria met in IMPLEMENTATION_PLAN.

+++ runtime/core/overflow.c
@@ -49,7 +49,10 @@ int16_t __dao_wrapping_sub_i16(int16_t a, int16_t b) {
 }
 
 int16_t __dao_wrapping_mul_i16(int16_t a, int16_t b) {
-  return u16_to_i16((uint16_t)a * (uint16_t)b);
+  // Cast to uint32_t before multiplying to avoid C integer-promotion UB:
+  // (uint16_t)a * (uint16_t)b promotes both operands to signed int on
+  // targets where int is 32-bit, and 65535*65535 overflows int.
+  return u16_to_i16((uint16_t)((uint32_t)(uint16_t)a * (uint32_t)(uint16_t)b));
 }
 
 int32_t __dao_wrapping_add_i32(int32_t a, int32_t b) {
